requirements.txt
anyio==3.5.0
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
docopt==0.6.2
fastapi==0.72.0
google-api-core==2.4.0
google-api-python-client==2.35.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
h11==0.12.0
httplib2==0.20.2
httptools==0.3.0
idna==3.3
oauthlib==3.1.1
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyparsing==3.0.6
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
